Added separate STL adapter, added std::string casts
